requirements.txt
####streamlit==1.21.0
streamlit==1.36.0
watchdog==4.0.1
joblib==1.3.2
numpy==1.24.3
pandas==1.5.3
##matplotlib==3.7.5
##seaborn==0.13.1
scikit-learn==1.3.2
lightgbm==4.1.0
##xgboost==1.7.6
pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
scikit-image==0.22.0